+++ src/core/pv.jl
@@ -33,24 +33,38 @@ struct PV <: AbstractTech
     module_type
     losses
     azimuth
+    gcr
+    radius
     name
     location
     existing_kw
     min_kw
     max_kw
-    cost_per_kw
+    installed_cost_per_kw
     om_cost_per_kw
     degradation_pct
     macrs_option_years
     macrs_bonus_pct
     macrs_itc_reduction
-    total_itc_pct
-    total_rebate_per_kw
     kw_per_square_foot
     acres_per_kw
     inv_eff
     dc_ac_ratio
     prod_factor_series_kw
+    federal_itc_pct
+    federal_rebate_per_kw
+    state_ibi_pct
+    state_ibi_max
+    state_rebate_per_kw
+    state_rebate_max
+    utility_ibi_pct
+    utility_ibi_max
+    utility_rebate_per_kw
+    utility_rebate_max
+    production_incentive_per_kwh
+    production_incentive_max_benefit
+    production_incentive_years
+    production_incentive_max_kw
 
     function PV(;
         tilt::Real,
@@ -58,24 +72,38 @@ struct PV <: AbstractTech
         module_type::Int=0,
         losses::Real=0.14,
         azimuth::Real=180,
+        gcr::Real=0.4,
+        radius::Int=0,
         name::String="PV",
         location::String="both",
         existing_kw::Real=0,
         min_kw::Real=0,
-        max_kw::Real=1.0e5,
-        cost_per_kw::Real=1600.0,
+        max_kw::Real=1.0e9,
+        installed_cost_per_kw::Real=1600.0,
         om_cost_per_kw::Real=16.0,
         degradation_pct::Real=0.005,
         macrs_option_years::Int = 5,
         macrs_bonus_pct::Float64 = 1.0,
         macrs_itc_reduction::Float64 = 0.5,
-        total_itc_pct::Float64 = 0.26,
-        total_rebate_per_kw::Float64 = 0.0,
         kw_per_square_foot::Float64=0.01,
         acres_per_kw::Float64=6e-3,
         inv_eff::Float64=0.96,
         dc_ac_ratio::Float64=1.2,
-        prod_factor_series_kw::Union{Missing, Array{Real,1}} = missing
+        prod_factor_series_kw::Union{Missing, Array{Real,1}} = missing,
+        federal_itc_pct::Float64 = 0.26,
+        federal_rebate_per_kw::Float64 = 0.0,
+        state_ibi_pct::Float64 = 0.0,
+        state_ibi_max::Float64 = 1.0e10,
+        state_rebate_per_kw::Float64 = 0.0,
+        state_rebate_max::Float64 = 1.0e10,
+        utility_ibi_pct::Float64 = 0.0,
+        utility_ibi_max::Float64 = 1.0e10,
+        utility_rebate_per_kw::Float64 = 0.0,
+        utility_rebate_max::Float64 = 1.0e10,
+        production_incentive_per_kwh::Float64 = 0.0,
+        production_incentive_max_benefit::Float64 = 1.0e9,
+        production_incentive_years::Float64 = 1.0,
+        production_incentive_max_kw::Float64 = 1.0e9
         )
 
         # validate inputs
@@ -118,24 +146,38 @@ struct PV <: AbstractTech
             module_type,
             losses,
             azimuth,
+            gcr,
+            radius,
             name,
             location,
             existing_kw,
             min_kw,
             max_kw,
-            cost_per_kw,
+            installed_cost_per_kw,
             om_cost_per_kw,
             degradation_pct,
             macrs_option_years,
             macrs_bonus_pct,
             macrs_itc_reduction,
-            total_itc_pct,
-            total_rebate_per_kw,
             kw_per_square_foot,
             acres_per_kw,
             inv_eff,
             dc_ac_ratio,
-            prod_factor_series_kw
+            prod_factor_series_kw,
+            federal_itc_pct,
+            federal_rebate_per_kw,
+            state_ibi_pct,
+            state_ibi_max,
+            state_rebate_per_kw,
+            state_rebate_max,
+            utility_ibi_pct,
+            utility_ibi_max,
+            utility_rebate_per_kw,
+            utility_rebate_max,
+            production_incentive_per_kwh,
+            production_incentive_max_benefit,
+            production_incentive_years,
+            production_incentive_max_kw
         )
     end
 end