+++ src/core/utils.jl
@@ -44,6 +44,22 @@ function annuity(years::Int, rate_escalation::Float64, rate_discount::Float64)
 end
 
 
+function annuity_escalation(analysis_period::Int, rate_escalation::Float64, rate_discount::Float64)
+    """
+    :param analysis_period: years
+    :param rate_escalation: escalation rate
+    :param rate_discount: discount rate
+    :return: present worth factor with escalation (inflation, or degradation if negative)
+    NOTE: assumes escalation/degradation starts in year 2 (unlike the `annuity` function above)
+    """
+    pwf = 0
+    for yr in range(1, stop=analysis_period + 1)
+        pwf += (1 + rate_escalation)^(yr - 1) / (1 + rate_discount)^yr
+    end
+    return pwf
+end
+
+
 function levelization_factor(years::Int, rate_escalation::Float64, rate_discount::Float64, 
     rate_degradation::Float64)
     #=
@@ -141,7 +157,13 @@ end
 function dictkeys_tosymbols(d::Dict)
     d2 = Dict()
     for (k, v) in d
-        if k in ["loads_kw", "prod_factor_series_kw"] && !isempty(v)
+        if k in [
+            "loads_kw", "critical_loads_kw",
+            "monthly_totals_kwh",
+            "prod_factor_series_kw", 
+            "monthly_energy_rates", "monthly_demand_rates",
+            "wholesale_rate"
+            ] && !isnothing(v)
             try
                 v = convert(Array{Real, 1}, v)
             catch
@@ -152,6 +174,7 @@ function dictkeys_tosymbols(d::Dict)
     end
     return d2
 end
+
 
 function filter_dict_to_match_struct_field_names(d::Dict, s::DataType)
     f = fieldnames(s)